Migrated AlbumsContainer to Styled Components
